Change header for CFG_UpdateConfigSavegame to increase userland accessibility
The current cfg:i version of this header (0x8030000) is only used by the 3DS in System Settings, while the cfg:s version (0x4030000) is used in several System Applications and System Modules, and is also compatible with System Settings.

Although a moot point in a Luma3DS-like environment, it can be a very convenient change for exploit homebrew apps in a stock userland environment. There is no downside to the change, and should not break any existing homebrew -- the functionality of CFG_UpdateConfigSavegame is the same with either header.